+++ OrderBook/OrderBook/Entities/Trade.swift
@@ -25,7 +25,7 @@ public struct TradeData: Decodable {
     let symbol: String
     let side: Side
     let size: UInt64
-    let price: Double
+    let price: Decimal
     let tickDirection: String
     let trdMatchID: String
     let grossValue: UInt64
